requirements.txt
# ── API Framework ────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Async Task Queue ─────────────────────────────────────────────
celery[redis]==5.3.6
redis==5.0.4

# ── Settings ─────────────────────────────────────────────────────
pydantic-settings==2.2.1

# ── Image Processing (Phase 3 — Week 3-4) ────────────────────────
rembg==2.0.56
Pillow==10.3.0
aiofiles==23.2.1
numpy<2                 # onnxruntime 1.x NumPy 2.x ABI ile uyumsuz
onnxruntime==1.19.2    # rembg backend (NumPy 1.x uyumlu)

# ── (Phase 2) Database ───────────────────────────────────────────
sqlalchemy==2.0.29      # ORM ve engine
psycopg2-binary==2.9.9  # PostgreSQL surucusu (binary, derleme gerektirmez)
# asyncpg==0.29.0       # Async surucusu (ileride gerekirse)
# alembic==1.13.1       # DB migration (ileride gerekirse)

# ── (Phase 3) AI / Storage ───────────────────────────────────────
boto3==1.34.0           # AWS S3 dosya depolama
httpx==0.27.0          # Tripo3D API HTTP istemcisi
# nudenet==3.4.2

